pyproject.toml
# This file contains project information and helps control the version numbers displayed.
# It also contains settings for linters and code checkers isort, black and mypy.
# Note that these settings are not respected with pre-commit run --all-files
# In that case add configurations to the .pre-commit-config.yaml file.

[tool.poetry]
name = "serve"
version = "1.0.0"
description = "SciLifeLab Serve is a platform offering machine learning model serving, app hosting (Shiny, Streamlit, Dash, etc.), and other tools to life science researchers affiliated with a Swedish research institute."
license = "Apache 2.0"
maintainers = ["Team Whale <serve@scilifelab.se>",]
authors = ["Team Whale <serve@scilifelab.se>",]
homepage = "https://serve.scilifelab.se"
repository = "https://github.com/ScilifelabDataCentre/stackn"
keywords = ["machine learning", "life science", "research", "django", "python"]
packages = []

[tool.poetry.dependencies]
python = "^3.10.0"

django = "==5.0.2"
django-celery-beat = "==2.6.0"
django-compressor = "==4.4"
django-cors-headers = "==4.3.1"
django-extensions = "==3.2.3"
django-filter = "==24.2"
django-tagulous = "==1.3.3"
django-guardian = "==2.4.0"
djangorestframework = "== 3.15.1"
django-crispy-forms = "==2.1"
crispy-bootstrap5 = "==2024.2"

# django-wiki
wiki = "==0.11.1"


# Other Python and project-related libraries
flower = "==2.0.1"
requests = "==2.31.0"
amqp = "==5.1.1"
psycopg2-binary = "==2.9.9"
redis = "==5.0.1"
watchdog = "==3.0.0"
drf-nested-routers = "==0.93.4"
s3fs = "==2023.9.2"
minio = "==7.1.17"
celery = "==5.3.4"
uvicorn = "==0.20.0"
flask-cors = "==4.0.0"
Pillow = "10.2.0"

flatten-json = "==0.1.13"
PyYAML = "==6.0.1"
Markdown = ">=3.4,<3.6"

# Logging
django-structlog = "^7.1.0"
colorlog = "^6.8.2"


black = { version = "==23.3.0", optional = true }
isort = { version = "==5.12.0", optional = true }
flake8 = { version = "==6.0.0", optional = true }
mypy = { version = "==1.10.0", optional = true }
#django-stubs = { version = "==5.0.0", optional = true }
hypothesis = { version = "==6.86.1", optional = true }
moto = { version = "==4.*", extras = ["ec2", "s3", "all"], optional = true }
pytest = { version = "~8.0.1", optional = true }
pytest-django = { version = "==4.8.0",  optional = true }
pytest-cov = { version = "==4.1.0", optional = true }
pytest-xdist = { version = "^3.5.0", optional = true }
django-axes = {extras = ["ipware"], version = "^6.3.0"}
django-password-validators = "^1.7.3"


[tool.poetry.extras]
linters = ["black", "isort", "flake8", "mypy"]
tests = ["pytest-django", "pytest-cov", "pytest-xdist", "hypothesis", "pytest", "moto"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.serve]
build-date = ""
gitref = ""
imagetag = ""

[tool.isort]
profile = 'black'

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.venv
  | venv
  | migrations
)/
'''

[tool.mypy]
strict = false
python_version = "3.12"
disallow_subclassing_any = false
ignore_missing_imports = false
warn_return_any = true
untyped_calls_exclude = ["apps.helpers","projects.models"]
exclude = ["venv", ".venv", "examples"]
#plugins = ["mypy_django_plugin.main"]

#[tool.django-stubs]
#django_settings_module = "studio.settings"

[[tool.mypy.overrides]]
module = "studio.*"
strict = true
disallow_subclassing_any = false
disallow_untyped_decorators = false
follow_imports_for_stubs = false

[[tool.mypy.overrides]]
module = "*.models"
strict = false
check_untyped_defs = false
disallow_incomplete_defs = false
disallow_subclassing_any = false
disallow_untyped_decorators = false
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = [
	"*.tests.*",
	"*.tests",
	"conftest",
	"cypress.*"
]
check_untyped_defs = false
disallow_incomplete_defs = false
disallow_subclassing_any = false
disallow_untyped_calls = false
disallow_untyped_decorators = false
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

# A temporary section to ease the transition of mypy into the codebase
[[tool.mypy.overrides]]
module = [
	"api.*",
	"apps.*",
	"collections_module.*",
	"common.*",
	"customtags.*",
	"models.*",
	"monitor.*",
	"news.*",
	"portal.*",
	"projects.*",
	"scripts.*",
]
strict = false
check_untyped_defs = false
disallow_incomplete_defs = false
disallow_subclassing_any = false
disallow_untyped_calls = false
disallow_untyped_decorators = false
disallow_untyped_defs = false
no_implicit_reexport = false

[[tool.mypy.overrides]]
module = [
	"boto3.*",
	"colorlog.*",
	"django.*",
	"django_structlog.*",
	"rest_framework.*",
	"flatten_json.*",
	"guardian.*",
	"hypothesis.*",
	"django_filters.*",
	"rest_framework_nested.*",
	"tagulous.*",
	"minio.*",
	"moto.*",
	"celery.*",
	"dash.*",
	"dash_core_components.*",
	"dash_html_components.*",
	"django_plotly_dash.*",
	"dash_bootstrap_components.*",
	"markdown.*",
	"pytest.*",
	"pytz.*",
	"requests.*",
	"s3fs.*",
	"setuptools.*",
	"structlog.*",
	"yaml.*",
	]
ignore_missing_imports = true
